requirements.txt
yfinance>=0.2
alpaca-trade-api>=3.0
fastapi>=0.100
uvicorn>=0.23
typer>=0.9
pyyaml>=6.0
pandas>=2.0
numpy>=1.24
pytest>=7.0
